SAA-264: Refactoring the location-related endpoints to group them.
