--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsactivitiesmanagementapi/resource/LocationPrefixController.kt
@@ -1,67 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsactivitiesmanagementapi.resource
-
-import io.swagger.v3.oas.annotations.Operation
-import io.swagger.v3.oas.annotations.media.Content
-import io.swagger.v3.oas.annotations.media.Schema
-import io.swagger.v3.oas.annotations.responses.ApiResponse
-import io.swagger.v3.oas.annotations.responses.ApiResponses
-import org.springframework.http.MediaType
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.ResponseBody
-import org.springframework.web.bind.annotation.RestController
-import uk.gov.justice.digital.hmpps.hmppsactivitiesmanagementapi.config.ErrorResponse
-import uk.gov.justice.digital.hmpps.hmppsactivitiesmanagementapi.model.whereabouts.LocationPrefixDto
-import uk.gov.justice.digital.hmpps.hmppsactivitiesmanagementapi.service.LocationService
-
-@RestController
-@RequestMapping("/prisons/{prisonCode}/location-prefix", produces = [MediaType.APPLICATION_JSON_VALUE])
-class LocationPrefixController(private val locationService: LocationService) {
-
-  @GetMapping
-  @ResponseBody
-  @Operation(
-    summary = "Get location prefix by group name",
-    description = "Get location prefix by group name",
-  )
-  @ApiResponses(
-    value = [
-      ApiResponse(
-        responseCode = "200",
-        description = "Successful call - Location prefix found",
-        content = [
-          Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = LocationPrefixDto::class)
-          )
-        ],
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Invalid request",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorised, requires a valid Oauth2 token",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Forbidden, requires an appropriate role",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Requested resource not found",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      )
-    ]
-  )
-  fun getLocationPrefix(
-    @PathVariable("prisonCode") prisonCode: String,
-    @RequestParam(value = "groupName", required = true) groupName: String,
-  ): LocationPrefixDto? = locationService.getLocationPrefixFromGroup(prisonCode, groupName)
-}
